Reorganiza testes de valores personalizados no arquivo de valores específicos
